Results/requirements.txt
# 🦷 Dental AI Requirements
# Core ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
detectron2

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0
albumentations>=1.3.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Utils
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0

# Development
jupyter>=1.0.0
ipywidgets>=8.0.0